refactor: merge the sprint game folder structure with the common structure
