+++ src/index.ts
@@ -1,6 +1,12 @@
 import App from './App';
 import './styles/main.scss';
 
+import './assets/images/sprint/icons/soundOn.svg';
+import './assets/images/sprint/icons/soundOff.svg';
+import './assets/images/sprint/icons/sound-on_121151.svg';
+import './assets/images/sprint/icons/home.svg';
+import './assets/images/sprint/icons/premium-icon-sound-3293595.png';
+
 const app = new App();
 document.addEventListener('DOMContentLoaded', () => {
   app.init();
