+++ src/styles/main.scss
@@ -16,5 +16,6 @@
 @import 'layout/minigames';
 @import 'layout/about';
 @import 'layout/footer';
+@import 'layout/sprint'
